factor out shared functionality in abstract learner to core
