+++ core/src/autogluon/core/learner/__init__.py
@@ -1,0 +1,3 @@
+from .abstract_learner import AbstractLearner
+
+__all__ = ["AbstractLearner"]
